pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vibescan-mcp-server"
version = "0.1.0"
description = "MCP server for VibeScan — scan projects for leaked secrets and security issues"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    { name = "Junsu Lee", email = "aguantar@users.noreply.github.com" },
]
keywords = [
    "mcp",
    "security",
    "secrets",
    "vibescan",
    "code-review",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
]
dependencies = [
    "mcp>=1.0.0",
    "vibescan-cli>=0.2.5",
]

[project.urls]
Homepage = "https://github.com/Aguantar/vibescan-mcp-server"
Repository = "https://github.com/Aguantar/vibescan-mcp-server"

[project.scripts]
vibescan-mcp-server = "vibescan_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/vibescan_mcp"]

[tool.hatch.metadata]
license-files = []
